Refactored tokenization service
