+++ handler/config_handler.py
@@ -2,7 +2,9 @@ from typing import Dict, Any, Optional, Union
 import requests
 import json
 import time
-import daemon
+
+from misc import daemon
+
 
 __config = None
