+++ misc/decorator.py
